+++ web/src/app/modules/eros-chat/directives.js
@@ -3,38 +3,30 @@
 /* Chat directives */
 
 angular.module('erosApp.chat', ['ngAudio'])
-.directive('erosChat', ['$rootScope', function($rootScope){
+.directive('erosChat', ['$rootScope','notifier', function($rootScope, notifier){
 	return {
 		restrict: 'A',
 		templateUrl: 'modules/eros-chat/chat.tpl.html',
 		controller: ['$scope',  function($scope){
+
 			$scope.selectRoom = function(room){
 				if(typeof room == "object"){
 					$scope.$parent.selectedRoom = room
-				}else{
-					$scope.$parent.selectedRoom = $scope.rooms[Object.keys($scope.rooms)[0]]
-					$scope.defaultRoom = $scope.$parent.selectedRoom
+					$rootScope.$emit("chat_room","selectedRoom")
 				}
-				$rootScope.$emit("chat_room","selectedRoom")
 			}
 
 			$scope.sendChatMessage = function(target, message) {
 				if(message[0] === "@"){
 					target = message.split(" ",1)[0].split("@")[1]
 					eros.chat.sendToPriv(target, message.split(" ").slice(1).join(" "))
+					$scope.selectRoom($scope.privs[target.toLowerCase()])
 				}else if(typeof ($scope.selectedRoom.priv) == 'object'){
 					eros.chat.sendToPriv($scope.selectedRoom.priv.name, message)
 				}else{
 					eros.chat.sendToRoom(target, message);
 				}
 				// $scope.chatMessage = "real"
-			}
-
-			$scope.setDefaultRoom = function(room){
-				if(typeof $scope.$parent.selectedRoom == 'undefined'){
-					$scope.selectRoom(room)
-					$scope.defaultRoom = room
-				}
 			}
 
 			$scope.addUserMsg = function(user){
@@ -47,8 +39,16 @@ angular.module('erosApp.chat', ['ngAudio'])
 				$scope.$parent.chatMessage = newmessage
 
 				document.getElementById("chat-input").childNodes[0].focus()
-		
-				
+			}
+
+			$scope.openPrivUser = function(username){
+				if(typeof username === 'string' && username != eros.localUser.username){
+					if(typeof $scope.privs[username.toLowerCase()] === 'undefined'){
+						eros.chat.joinPriv(username)
+					}
+
+					$scope.selectRoom($scope.privs[username.toLowerCase()])
+				}
 			}
 
 			$scope.updateChatInput = function(message){