+++ web/CHANGELOG.md
@@ -4,6 +4,28 @@
 - Make config for local development (mocking, etc.)
 - Join private Room (just UI)
 - Create Room (just UI)
+- Draw
+- Matched notification volume
+- Lots of error server error catching
+
+## 0.1.0 - 2014-11-30
+### Added 
+- Matchmaking system
+- Page title notifications
+- Matched notification
+- Authtoken field as password
+- Version/feedback footer
+
+### Changed
+- Private messaging now needs double click (simple click on user does nothing)
+- Default channel now works
+- Moved sound notification volume to user list
+- Stats moved to bottom
+
+## 0.0.6 - 2014-09-17
+### Added
+- Sound noitifications
+- [T] Initial matchmaking communications
 
 ## 0.0.5 - 2014-09-15
 ### Added 
@@ -24,10 +46,3 @@
 - Change header font, as it won't work due to lack of all font files.
 - Adapt rank to body font
 - Improved channel closing functionality
-
-
-
-
-
-
-